+++ lib/lightning/extensions/run_queue.ex
@@ -5,10 +5,14 @@ defmodule Lightning.Extensions.RunQueue do
 
   @callback enqueue(
               run ::
-                Lightning.Run.t() | Ecto.Changeset.t(Lightning.Run.t())
+                Lightning.Run.t()
+                | Ecto.Changeset.t(Lightning.Run.t())
+                | Ecto.Multi.t()
             ) ::
               {:ok, Lightning.Run.t()}
               | {:error, Ecto.Changeset.t(Lightning.Run.t())}
+              | {:error, Ecto.Multi.name(), any(),
+                 %{required(Ecto.Multi.name()) => any()}}
 
   @callback claim(demand :: non_neg_integer()) ::
               {:ok, [Lightning.Run.t()]}